changed to window size again
The previous "fix" for window size was wrong because of another issue that was responsible for the different window size on Windows. That underlying issue is now fixed and this "band-aid" is now removed. The code is not OS specific anymore...for later releases an OS specific implementation might be necessary if further kinds of OS should be supported